requirements.txt
SQLAlchemy==2.0.44
fastapi==0.123.4
pydantic==2.4.2
uvicorn==0.22.0
db-sqlite3==0.0.1
psycopg2-binary
python-dotenv
numpy
google-generativeai 